+++ src/main/java/com/driply/payments/payment/dto/PaymentResultDTO.java
@@ -1,7 +1,5 @@
 package com.driply.payments.payment.dto;
 
-import com.driply.payments.payment.entity.PaymentStatus;
-
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
