+++ src/main/java/com/driply/payments/payment/dto/PaymentStatus.java
@@ -1,0 +1,9 @@
+package com.driply.payments.payment.dto;
+
+public enum PaymentStatus {
+	PAID,
+	PENDING,
+	FAILED,
+	CANCELED,
+	REFUNDED
+}
