--- src/main/java/com/driply/payments/payment/entity/PaymentError.java
@@ -1,14 +1,0 @@
-package com.driply.payments.payment.entity;
-
-import java.time.LocalDateTime;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class PaymentError {
-    private String errorCode;
-    private String message;
-    private LocalDateTime timestamp;
-}
