--- src/main/java/com/driply/payments/payment/entity/PaymentStatus.java
@@ -1,9 +1,0 @@
-package com.driply.payments.payment.entity;
-
-public enum PaymentStatus {
-    PAID,
-    PENDING,
-    FAILED,
-    CANCELED,
-    REFUNDED
-}
