+++ src/test/java/com/driply/payments/payment/gateway/PaymentGatewayFactoryTest.java
@@ -9,31 +9,31 @@ import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.web.reactive.function.client.WebClient;
 
-import com.driply.payments.payment.entity.PGType;
+import com.driply.payments.payment.dto.PGType;
 
 @SpringBootTest
 class PaymentGatewayFactoryTest {
 
-    private PaymentGatewayFactory paymentGatewayFactory;
-    private WebClient tossWebClient = WebClient.builder()
-        .baseUrl("https://api.tosspayments.com")
-        .build();
+	private PaymentGatewayFactory paymentGatewayFactory;
+	private WebClient tossWebClient = WebClient.builder()
+		.baseUrl("https://api.tosspayments.com")
+		.build();
 
-    @BeforeEach
-    void setUp() {
-        List<PaymentGateway> gateways = List.of(new TossPaymentsGateway(tossWebClient));
-        paymentGatewayFactory = new PaymentGatewayFactory(gateways);
-    }
+	@BeforeEach
+	void setUp() {
+		List<PaymentGateway> gateways = List.of(new TossPaymentsGateway(tossWebClient));
+		paymentGatewayFactory = new PaymentGatewayFactory(gateways);
+	}
 
-    @Test
-    @DisplayName("pgType이 TOSS일 때, TossPaymentGateway를 반환")
-    void getGateway() {
-        String pgType = PGType.TOSS.name();
-        PaymentGateway gateway = paymentGatewayFactory.getGateway(pgType);
+	@Test
+	@DisplayName("pgType이 TOSS일 때, TossPaymentGateway를 반환")
+	void getGateway() {
+		String pgType = PGType.TOSS.name();
+		PaymentGateway gateway = paymentGatewayFactory.getGateway(pgType);
 
-        Assertions.assertNotNull(gateway);
-        Assertions.assertInstanceOf(TossPaymentsGateway.class, gateway);
-        TossPaymentsGateway tossPaymentsGateway = (TossPaymentsGateway) gateway;
-        Assertions.assertEquals(pgType, tossPaymentsGateway.getPGType());
-    }
+		Assertions.assertNotNull(gateway);
+		Assertions.assertInstanceOf(TossPaymentsGateway.class, gateway);
+		TossPaymentsGateway tossPaymentsGateway = (TossPaymentsGateway)gateway;
+		Assertions.assertEquals(pgType, tossPaymentsGateway.getPGType());
+	}
 }
